+++ types/index.d.ts
@@ -1,11 +1,12 @@
 type TDate = Date | String | Number;
+type TMap = { [k: string]: unknown };
 
 export const time: {
   /**
    * Compare any Dates
    *  Create any dates compare functions or use our presets.
    * @example
-   * const astropack = require('astro-kit');
+   * const astropack = require('astropack');
    * astropack.time.compare((a, b) => a > b)('2023-05-07', '2023-05-08'); // false
    * astropack.time.compare((a, b) => a > b)('2023-01-01', '2021-05-08'); // true
    * astropack.time.compare.bigger('2023-01-01', '2021-05-08'); // true
@@ -26,7 +27,7 @@ export const time: {
    * astropack.time.prettify('h:m:s / D.M|Y', new Date()); // 18:50:54 / 07.05|2023
    * astropack.time.prettify('h-m-s.i', new Date()); // 18-50-54.045
    */
-  prettify: (format: string, date: Date | number | string) => string;
+  prettify: (format?: string, date?: Date | number | string) => string;
   /**
    * Time formatter
    * Make millisecond's more readable with format;
@@ -76,4 +77,179 @@ export const time: {
    * astropack.time.duration('1d 1h 1m 5s'); // 90065000
    */
   duration: (time: string) => number;
+};
+
+export const fs: {
+  file: {
+    /**
+     * Ensure that dir exists
+     *  Will create a directory if it doesn't exist
+     * @example
+     * const astropack = require('astropack');
+     * astropack.fs.file.name('/path/to/file.ext'); // file
+     */
+    name: (name: string) => string;
+    /**
+     * Ensure that dir exists
+     *  Will create a directory if it doesn't exist
+     * @example
+     * const astropack = require('astropack');
+     * astropack.fs.file.ext('/path/to/file.ext'); // ext
+     */
+    ext: (name: string) => string;
+    /**
+     * Ensure that dir exists
+     *  Will create a directory if it doesn't exist
+     * @example
+     * const astropack = require('astropack');
+     * astropack.fs.file.dir('/path/to/file.ext'); // /path/to
+     */
+    dir: (name: string) => string;
+  };
+  dir: {
+    /**
+     * Ensure that dir exists
+     *  Will create a directory if it doesn't exist
+     * @example
+     * const astropack = require('astropack');
+     * astropack.fs.dir.ensure('/path/to/dir').then(status => console.log(status));
+     */
+    ensure: (path: string) => boolean;
+    /**
+     * Ensure that dir exists
+     * @example
+     * const astropack = require('astropack');
+     * astropack.fs.dir.check('/path/to/dir').then(status => console.log(status));
+     */
+    check: (path: string) => boolean;
+  };
+};
+
+export const utils: {
+  /**
+   * Equality check
+   * @example
+   * const astropack = require('astropack');
+   * astropack.utils.equals(1, 1); // true
+   * astropack.utils.equals('test', 'test'); // true
+   * astropack.utils.equals({ a: 1 }, { a: 1 }); // true
+   * astropack.utils.equals([1, 2, [3]], [1, 2, [3]]); // true
+   * astropack.utils.equals({ a: { b: 'c' } }, { a: { b: 'c' } }); // true
+   */
+  equals: <T>(a: T, b: T) => boolean;
+  /**
+   * Equality check
+   * Can contain up to YB
+   * @example
+   * const astropack = require('astropack');
+   * astropack.utils.prettyBytes(1100); // '1.1 KB'
+   */
+  prettyBytes: (bytes: number) => string;
+};
+
+export const string: {
+  /**
+   * Parse value from string
+   * @example
+   * const astropack = require('astropack');
+   * astropack.string.from('false'); // false
+   * astropack.string.from('undefined'); // undefined
+   * astropack.string.from('null'); // null
+   * astropack.string.from('12.23'); // 12.23
+   * astropack.string.from('test'); // 'test'
+   * astropack.string.from('{"test":true}'); // { test: true }
+   */
+  from: (value: string) => unknown;
+  /**
+   * Parse value to string
+   * @example
+   * const astropack = require('astropack');
+   * astropack.string.to(false); // 'false'
+   * astropack.string.to(undefined); // 'undefined'
+   * astropack.string.to('12.23'); // '12.23'
+   * astropack.string.to('test'); // 'test'
+   * astropack.string.to({ test: true }); // '{"test":true}'
+   */
+  to: (value: unknown) => string;
+  /**
+   * Safe json parser, returns null if error
+   * @example
+   * const astropack = require('astropack');
+   * astropack.string.jsonParse('{'); // null
+   * astropack.string.jsonParse('{"test":true}'); // { "test": true }
+   */
+  jsonParse: (value: unknown) => object | null;
+  /**
+   * String template creator
+   * @example
+   * const astropack = require('astropack');
+   * const temp = astropack.string.template`Hello ${'user'} !`;
+   * temp({ user: 'astro' }); // 'Hello astro !';
+   */
+  template: (s: string[], ...keys: string[]) => (values: TMap) => string;
+  case: {
+    /**
+     * @example
+     * const astropack = require('astropack');
+     * const temp = astropack.string.case.isConstant('GLOBAL'); // true
+     */
+    isConstant: (s: string) => boolean;
+    /**
+     * @example
+     * const astropack = require('astropack');
+     * const temp = astropack.string.case.isFirstUpper('Hello'); // true
+     */
+    isFirstUpper: (s: string) => boolean;
+    /**
+     * @example
+     * const astropack = require('astropack');
+     * let temp = astropack.string.case.isFirstLetter('H1'); // true
+     * temp = astropack.string.case.isFirstLetter('1H'); // false
+     */
+    isFirstLetter: (s: string) => boolean;
+    /**
+     * @example
+     * const astropack = require('astropack');
+     * let temp = astropack.string.case.isFirstLower('hello'); // true
+     */
+    isFirstLower: (s: string) => boolean;
+    /**
+     * @example
+     * const astropack = require('astropack');
+     * let temp = astropack.string.case.toCamel('-')('hello-world'); // helloWorld
+     * let temp = astropack.string.case.toCamel(' ')('hello world'); // helloWorld
+     */
+    toCamel: (sep: string) => (s: string) => string;
+    /**
+     * @example
+     * const astropack = require('astropack');
+     * let temp = astropack.string.case.fromCamel('-')('helloWorld'); // hello-world
+     * let temp = astropack.string.case.fromCamel(' ')('helloWorld'); // hello world
+     */
+    fromCamel: (sep: string) => (s: string) => string;
+    /**
+     * @example
+     * const astropack = require('astropack');
+     * let temp = astropack.string.case.spinalToCamel('hello-world'); // helloWorld
+     */
+    spinalToCamel: (s: string) => string;
+    /**
+     * @example
+     * const astropack = require('astropack');
+     * let temp = astropack.string.case.camelToSpinal('helloWorld'); // hello-world
+     */
+    camelToSpinal: (s: string) => string;
+    /**
+     * @example
+     * const astropack = require('astropack');
+     * let temp = astropack.string.case.camelToSnake('helloWorld'); // hello_world
+     */
+    camelToSnake: (s: string) => string;
+    /**
+     * @example
+     * const astropack = require('astropack');
+     * let temp = astropack.string.case.snakeToCamel('hello_world'); // helloWorld
+     */
+    snakeToCamel: (s: string) => string;
+  };
 };